# Conflicts:
#	happiness-backend/api/models/models.py

+++ happiness-backend/api/routes/user.py
@@ -9,12 +9,14 @@ from flask import Blueprint
 from flask import current_app
 
 from api.app import db
-from api.authentication.email_token_methods import confirm_email_token
-from api.dao import users_dao
-from api.models.models import User, Setting
+from api.authentication.auth import token_current_user
+from api.dao.users_dao import get_user_by_email
+from api.util.jwt_methods import verify_token
+from api.dao import users_dao, happiness_dao
+from api.models.models import User, Setting, Happiness, Journal
 from api.models.schema import UserSchema, CreateUserSchema, SettingsSchema, SettingInfoSchema, \
-    UserInfoSchema, PasswordResetReqSchema, SimpleUserSchema, EmptySchema, PasswordResetSchema, \
-    FileUploadSchema, GroupSchema, PasswordKeyOptSchema
+    UserInfoSchema, EmailSchema, SimpleUserSchema, EmptySchema, PasswordResetSchema, \
+    FileUploadSchema, AmountSchema, CountSchema, UserDeleteSchema
 from api.routes.token import token_auth
 from api.util import email_methods
 from api.util.errors import failure_response
@@ -38,10 +40,10 @@ def create_user(req):
 
     similar_user = users_dao.get_user_by_email(email)
     if similar_user is not None:
-        return failure_response("Provided data already exists", 400)
+        return failure_response("Provided email is already taken. Please try a different email.", 400)
     similar_user2 = users_dao.get_user_by_username(username)
     if similar_user2 is not None:
-        return failure_response("Provide data already exists", 400)
+        return failure_response("Provided username is already taken. Please try a different username.", 400)
     current_user = User(email=email, password=password, username=username)
     db.session.add(current_user)
     db.session.commit()
@@ -60,7 +62,7 @@ def get_user_by_id(user_id):
     User must share a group with the user they are viewing. \n
     Returns: JSON of User object containing user information
     """
-    current_user = token_auth.current_user()
+    current_user = token_current_user()
 
     friend_user = users_dao.get_user_by_id(user_id)
     if friend_user is None:
@@ -87,28 +89,30 @@ def get_user_by_username(username):
     return user_lookup
 
 
-@user.get('/groups')
-@authenticate(token_auth)
-@response(GroupSchema(many=True))
-def user_groups():
-    """
-    Get Groups
-    Returns: a list of happiness groups that the user is in as well as any they have been invited to join.
-    """
-    return token_auth.current_user().groups
-
-
 @user.delete('/')
+@body(UserDeleteSchema)
 @authenticate(token_auth)
-def delete_user():
+def delete_user(req):
     """
     Delete User
     Deletes the user that is currently logged in, including all user data.
+    Requires that the user inputs their password before deleting their account.
     """
-    current_user = token_auth.current_user()
+    if not token_current_user().verify_password(req.get("password")):
+        return failure_response("Incorrect Password", 401)
+
+    current_user = token_current_user()
+
+    happiness_records = db.session.query(Happiness).filter_by(user_id=current_user.id).all()
+    for happiness_record in happiness_records:
+        db.session.delete(happiness_record)
+
+    journal_records = db.session.query(Journal).filter_by(user_id=current_user.id).all()
+    for journal_record in journal_records:
+        db.session.delete(journal_record)
+
     db.session.delete(current_user)
     db.session.commit()
-
     return '', 204
 
 
@@ -124,7 +128,7 @@ def add_user_setting(req):
     If the setting already exists in the property bag, it can enable or disable the setting. \n
     Returns: A JSON success response that contains the added setting, or a failure response.
     """
-    current_user = token_auth.current_user()
+    current_user = token_current_user()
     key, enabled, value = req.get("key"), req.get("enabled"), req.get("value")
     old_setting = Setting.query.filter(
         Setting.user_id == current_user.id, Setting.key == key).first()
@@ -155,7 +159,7 @@ def get_user_settings():
     Gets the settings of the current user by authorization token. \n
     Returns: A JSON response of a list of keys, booleans, and values that contain setting keys, whether they are enabled/disabled, and specific values for the user.
     """
-    current_user = token_auth.current_user()
+    current_user = token_current_user()
     settings = Setting.query.filter(Setting.user_id == current_user.id).all()
     return settings
 
@@ -163,23 +167,23 @@ def get_user_settings():
 @user.put('/info/')
 @authenticate(token_auth)
 @body(UserInfoSchema)
-@arguments(PasswordKeyOptSchema, location='headers')
-@response(SimpleUserSchema, headers=PasswordKeyOptSchema)
+@response(SimpleUserSchema)
 @other_responses({400: "Provided data already exists or empty data field."})
-def change_user_info(req, headers):
+def change_user_info(req):
     """
     Change User Info
     Changes a user's info based on 3 different `data_type`(s):
     - "username" (must be unique)
     - "email" (must be unique)
-    - "password"
-    - "key_recovery_phrase"
+    - "password" (requires old password)
+    - "key_recovery_phrase" (requires current password)
     Then the associated data must be put in the `data` field of the request. \n
-    If changing password or adding/changing key_recovery_phrase, the user's `password_key`
-    (provided by server during API token creation) must also be sent.
+    If changing password, put the user's old password in the `data` field and the new password in the `data2` field. \n
+    If adding a password key recovery phrase, put the user's current password in the `data` field
+    and the recovery phrase in the `data2` field.
     """
     data_type = req.get("data_type")
-    current_user = token_auth.current_user()
+    current_user = token_current_user()
 
     if len(req.get("data")) == 0:
         return failure_response('Data is empty.', 400)
@@ -192,8 +196,6 @@ def change_user_info(req, headers):
             return failure_response("Provide data already exists", 400)
 
         current_user.username = new_username
-        db.session.commit()
-        return current_user
     elif data_type == "email":
         # Changes a user's email, which requires their email to be unique.
         new_email = req.get("data")
@@ -202,80 +204,71 @@ def change_user_info(req, headers):
             return failure_response("Provided data already exists", 400)
 
         current_user.email = new_email
-        db.session.commit()
-        return current_user
     elif data_type == "password":
-        # Changes a user's password.
-        try:
-            current_user.change_password(req.get("data"), headers.get("password_key"))
-            db.session.commit()
-            return current_user, {
-                'Password-Key': current_user.derive_pwd_key(req.get('data'))
-            }
-        except Exception as e:
-            print(e)
-            return failure_response('Invalid password key.', 400)
+        # Changes a user's password, which requires their old password and a new one.
+        old_password, new_password = req.get("data"), req.get("data2")
+        if not current_user.verify_password(old_password):
+            return failure_response("Incorrect Password", 401)
+
+        current_user.change_password(old_password, new_password)
     elif data_type == "key_recovery_phrase":
-        # Adds/Changes key recovery phrase to prevent loss of encrypted data on password reset.
-        try:
-            current_user.add_key_recovery(req.get("data"), headers.get("password_key"))
-            db.session.commit()
-            return current_user
-        except Exception as e:
-            print(e)
-            return failure_response('Invalid password key.', 400)
+        # Adds/Changes key recovery phrase to prevent loss of encrypted data on password reset,
+        # which requires their password and a recovery phrase.
+        password, recovery_phrase = req.get("data"), req.get("data2")
+        if not current_user.verify_password(password):
+            return failure_response("Incorrect Password", 401)
+
+        current_user.add_key_recovery(password, recovery_phrase)
     else:
         return failure_response('Unknown data_type.', 400)
+
+    db.session.commit()
+    return current_user
 
 
 @user.post('/reset_password/<token>')
 @body(PasswordResetSchema)
-@arguments(PasswordKeyOptSchema, location='headers')
 @response(EmptySchema, 204, 'Password reset successful')
 @other_responses({400: "Invalid password reset token or recovery input"})
-def reset_password(req, headers, token):
+def reset_password(req, token):
     """
     Reset Password from Token
-    This function was written under the assumption that when the user receives a verify password email,
-    they would be redirected to a page where they are prompted to enter a new password. Then from this page they
-    make a post request to the backend with their new intended password. \n
-
-    If the user has previously added a key recovery phrase and provides their recovery phrase, their
-    encrypted journal entries will not be lost. Otherwise, this will cause them to become lost
+    Resets a user's password using the JWT token which was sent to their email address. \n
+    If a user has encrypted journal entries, they will need to provide their key recovery phrase
+    in order to encrypt their journal entries using their new password. If they did not set up
+    a recovery key or do not provide one, the journal entries will be lost
     (since they will never be able to be decrypted) and therefore will be deleted.
     """
     # Verify token is not expired
-    email = confirm_email_token(token)
-    if email is False:
-        return failure_response("Token expired", 400)
-    current_user = users_dao.get_user_by_email(email)
+    token = verify_token(token)
+    if not token:
+        return failure_response("Invalid/Expired token", 400)
+    current_user = users_dao.get_user_by_email(token['reset_email'])
     if not current_user:
-        return failure_response("Password reset token verification failed, token may be expired", 401)
+        return failure_response("Password reset token verification failed", 401)
 
     try:
-        current_user.reset_password(req.get("password"),
-                                    req.get("recovery_phrase"), headers.get('password_key'))
+        current_user.reset_password(req.get("password"), recovery_phrase=req.get("recovery_phrase"))
         db.session.commit()
         return '', 204
     except Exception as e:
         print(e)
-        return failure_response('Invalid recovery input.', 400)
+        return failure_response('Invalid recovery phrase.', 400)
 
 
 @user.post('/initiate_password_reset/')
-@body(PasswordResetReqSchema)
+@body(EmailSchema)
 @response(EmptySchema, 204, 'Password reset email sent')
-@other_responses({404: "User associated with email address not found"})
+@other_responses({400: "User associated with email address not found"})
 def send_reset_password_email(req):
     """
     Send Reset Password Email
-    Sends a password reset request email to email sent in the req of the JSON request. \n
+    Sends a password reset request email to email provided in the JSON request. \n
     Returns: a success response or failure response depending on the result of the operation
     """
-    email = req.get("email")
-    user_by_email = users_dao.get_user_by_email(email)
+    user_by_email = users_dao.get_user_by_email(req.get("email"))
     if user_by_email is None:
-        return failure_response("User associated with email address not found", 404)
+        return failure_response("User associated with email address not found", 400)
     threading.Thread(target=email_methods.send_password_reset_email,
                      args=(user_by_email,)).start()
     return '', 204
@@ -289,7 +282,7 @@ def get_self():
     Get Self
     Returns: the user object corresponding to the currently logged-in user.
     """
-    return token_auth.current_user()
+    return token_current_user()
 
 
 @user.post('/pfp/')
@@ -308,7 +301,7 @@ def add_pfp(req):
     # Check valid user and valid image file
 
     data = req["file"].read()
-    current_user = token_auth.current_user()
+    current_user = token_current_user()
     # Check that data exists
     if not data:
         return failure_response("Invalid request", 400)
@@ -342,3 +335,36 @@ def add_pfp(req):
     db.session.commit()
 
     return current_user
+
+
+@user.get('/count/')
+@authenticate(token_auth)
+@arguments(AmountSchema)
+@response(CountSchema)
+def profile_stats(req):
+    """
+    Profile Stats
+    Returns the number of happiness entries that the user has made on happiness app and the number of groups that
+    the user is in.
+    """
+    user_id = req.get("user_id", token_current_user().id)
+    if not (user_id == token_current_user().id or
+            token_current_user().has_mutual_group(users_dao.get_user_by_id(user_id))):
+        return failure_response("Not Allowed.", 403)
+    num_groups = token_current_user().groups.count()
+    return {"entries": happiness_dao.get_num_of_entries(user_id, low=0, high=10), "groups": num_groups}
+
+
+@user.post('/nudge/')
+@authenticate(token_auth)
+@body(EmailSchema)
+def nudge_user(req):
+    """
+    Nudge User
+    Send an email to invite a non-registered user to create an account
+    """
+    if not get_user_by_email(req.get("email")):
+        threading.Thread(target=email_methods.send_nudge_email,
+                         args=(req.get("email"), token_current_user())).start()
+        return "", 204
+    return failure_response("User already exists.", 400)
